Refactor Dashboard.jsx
Add New icon to page
