+++ src/utils/helpers.js
@@ -1,9 +1,9 @@
 const userHasCoordinates = (user) => {
   if (
-    user?.current_coordinates?.coordinates[0] === null &&
-    user?.current_coordinates?.coordinates[1] === null &&
-    user?.city?.coordinates[0] === null &&
-    user?.city?.coordinates[1] === null
+    user?.current_coordinates?.coordinates[0] === 0 &&
+    user?.current_coordinates?.coordinates[1] === 0 &&
+    user?.city?.coordinates[0] === 0 &&
+    user?.city?.coordinates[1] === 0
   ) {
     return false;
   }
